web_server/requirements.txt

Flask==2.3.2
Flask-SQLAlchemy==3.0.3
Flask-Login==0.6.2
Werkzeug==2.3.6
python-dotenv==1.0.0
argon2-cffi==21.3.0
psycopg2-binary==2.9.6
flask-migrate==4.0.4

Authlib==1.2.1
requests==2.31.0
PyJWT==2.8.0
